fix/#478: 이미지 cdn 주소 받아오기 로직 모듈화
