+++ components/editor/TextEditor.tsx
@@ -6,6 +6,7 @@ import { useParams } from 'next/navigation';
 import styled from 'styled-components';
 
 import { postExternalImage } from '@/api/postImage';
+import { getChangedImgSrc } from '@/utils/getChangedImgSrc';
 import { getContentCtrlVImage } from '@/utils/getImageMultipartData';
 
 interface editorProps {
@@ -42,26 +43,8 @@ const TextEditor = ({ editor, handleDragOver, handleDrop }: editorProps) => {
     const targetImg = event.currentTarget.querySelector(
       'img:not([src^="https://cdn.palms.blog/"]):not([class^="ProseMirror"]):not([class="inaccessible"]), img[class="ProseMirror-selectednode"]:not([src^="https://cdn.palms.blog/"]):not([class^="inaccessible"])',
     );
-    // src뽑아내기
-    let imgSrc = targetImg?.getAttribute('src');
-    if (!imgSrc) return;
-
-    // 외부 링크인 경우 base64로 받아내기
-    if (!imgSrc.startsWith('data:')) {
-      const data = await postExternalImage(imgSrc);
-      // 접근 권한이 없는 경우
-      if (!data) {
-        targetImg?.classList.add('inaccessible');
-        return;
-      }
-      imgSrc = `data:image/png;base64,${data}`;
-    }
-
-    // base64 -> File -> CDN 주소 받아오기
-    const blob = await fetch(imgSrc).then((res) => res.blob());
-    const file = new File([blob], 'image.png', { type: 'image/png' });
-    const imgUrl = await getContentCtrlVImage(file, String(team));
-    if (imgUrl) targetImg?.setAttribute('src', imgUrl);
+    if (!targetImg) return;
+    await getChangedImgSrc(targetImg, String(team));
   };
 
   return (
